+++ backend/api/shopping_cart/shopping_cart/delete.yml
