+++ backend/app.py
@@ -8,7 +8,7 @@ from flask import Flask
 from auth.route import auth_bp
 from database import create_db_command, db
 from item.route import item_bp
-from shoppingcart.route import shoppingcart_bp
+from shopping_cart.route import shopping_cart_bp
 from static.route import static_bp
 from tag.route import tag_bp
 from user.route import user_bp
@@ -35,7 +35,7 @@ def create_app(test_config: Mapping[str, Any] | None = None) -> Flask:
     app.register_blueprint(static_bp)
     app.register_blueprint(tag_bp)
     app.register_blueprint(user_bp)
-    app.register_blueprint(shoppingcart_bp)
+    app.register_blueprint(shopping_cart_bp)
 
     @app.route("/", methods=["GET"])
     def index() -> str:
